switched css to tailwind, add todos, moved paymentForm to organisms
