+++ frontend/src/components/organisms/PaymentForm.vue
@@ -1,8 +1,8 @@
 <template>
-  <div class="payment-container">
-    <h2 class="payment-title">Pay Invoice</h2>
+  <div class="flex flex-col items-center justify-center max-w-sm mx-auto my-8 p-8 border border-gray-300 rounded-lg bg-white shadow-md">
+    <h2 class="mb-4 text-lg font-semibold">Pay Invoice</h2>
 
-    <div class="payment-details">
+    <div class="w-full space-y-4">
       <!-- Payment method section -->
       <div class="payment-method">
         <!-- Card icons here -->
@@ -12,46 +12,47 @@
       <div class="payment-amount">
         <label for="amount" class="block text-sm font-medium text-gray-700">Payment amount</label>
         <div class="mt-1 relative rounded-md shadow-sm">
-          <input type="text" id="amount" class="form-input block w-full sm:text-sm border-gray-300" placeholder="Canadian Dollars" readonly />
+          <input type="text" id="amount" class="form-input block w-full sm:text-sm border-gray-300 rounded-md p-2" placeholder="Canadian Dollars" readonly />
         </div>
       </div>
 
       <!-- Name on card section -->
       <div class="input-group">
-        <label for="cardName">Name on card</label>
-        <input type="text" id="cardName" v-model="cardName" class="form-input" placeholder="Name on card">
+        <label for="cardName" class="block mb-2">Name on card</label>
+        <input type="text" id="cardName" v-model="cardName" class="form-input block w-full p-2 border border-gray-300 rounded-md" placeholder="Name on card">
       </div>
 
       <!-- Card number section -->
       <div class="input-group">
-        <label for="cardNumber">Card number</label>
-        <input type="text" id="cardNumber" v-model="cardNumber" class="form-input" placeholder="Card number" maxlength="16">
+        <label for="cardNumber" class="block mb-2">Card number</label>
+        <input type="text" id="cardNumber" v-model="cardNumber" class="form-input block w-full p-2 border border-gray-300 rounded-md" placeholder="Card number" maxlength="16">
       </div>
 
       <!-- Expiry date and security code section -->
-      <div class="flex">
+      <div class="flex gap-4">
         <div class="input-group flex-1">
-          <label for="expiry">Expiry date</label>
-          <input type="text" id="expiry" v-model="expiry" class="form-input" placeholder="MM/YY">
+          <label for="expiry" class="block mb-2">Expiry date</label>
+          <input type="text" id="expiry" v-model="expiry" class="form-input block w-full p-2 border border-gray-300 rounded-md" placeholder="MM/YY">
         </div>
         <div class="input-group flex-1">
-          <label for="securityCode">Security code</label>
-          <input type="text" id="securityCode" v-model="securityCode" class="form-input" placeholder="CVV" maxlength="3">
+          <label for="securityCode" class="block mb-2">Security code</label>
+          <input type="text" id="securityCode" v-model="securityCode" class="form-input block w-full p-2 border border-gray-300 rounded-md" placeholder="CVV" maxlength="3">
         </div>
       </div>
 
       <!-- ZIP/Postal code section -->
       <div class="input-group">
-        <label for="zip">ZIP/Postal code</label>
-        <input type="text" id="zip" v-model="postalCode" class="form-input" placeholder="ZIP Code">
+        <label for="zip" class="block mb-2">ZIP/Postal code</label>
+        <input type="text" id="zip" v-model="postalCode" class="form-input block w-full p-2 border border-gray-300 rounded-md" placeholder="ZIP Code">
       </div>
 
       <!-- PrimeVue Button for payment submission -->
-      <Button label="Pay" @click="submitPayment" class="w-full mt-2" />
+      <Button label="Pay" @click="submitPayment" class="w-full mt-4" />
     </div>
     <Toast/>
   </div>
 </template>
+
 
 <script setup lang="ts">
 import { ref } from 'vue';
@@ -82,55 +83,7 @@ function submitPayment() {
   }
 
   toast.add({ severity: 'success', summary: 'Payment Successful', detail: 'Payment processed successfully', life: 3000 });
-  // Implement payment processing logic here...
+  //TODO: Implement payment processing logic here...
 }
 
 </script>
-
-<style scoped>
-.payment-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  max-width: 400px;
-  margin: 2rem auto;
-  padding: 2rem;
-  border: 1px solid #ccc;
-  border-radius: 0.5rem;
-  background: #fff;
-  box-shadow: 0 4px 8px rgba(0,0,0,0.1);
-}
-
-.payment-title {
-  margin-bottom: 1rem;
-}
-
-.payment-method, .payment-amount, .input-group {
-  width: 100%;
-  margin-bottom: 1rem;
-}
-
-.form-input {
-  padding: 0.5rem;
-  border: 1px solid #ccc;
-  border-radius: 0.375rem;
-  width: 100%;
-}
-
-  .flex {
-    display: flex;
-    gap: 1rem;
-  }
-
-  .flex-1 {
-    flex: 1;
-  }
-
-  .input-group label {
-    display: block;
-    margin-bottom: 0.5rem;
-  }
-
-  /* Import PrimeVue button styles if not globally imported */
-  </style>
