+++ frontend/src/views/PaymentView.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import Payment from '../components/PaymentForm.vue'
+import Payment from '../components/organisms/PaymentForm.vue'
 </script>
 
 <template>
